+++ src/test/converter2/issues/gh2681.ts
@@ -1,4 +1,0 @@
-/**
- * {@link Generator}
- */
-export const bug = 123;